requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bzip2==1.0.8
ca-certificates==2024.3.11
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.7
libffi==3.4.4
ncurses==6.4
openssl==3.0.13
pip==23.3.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.0
pydantic-core==2.18.1
pyparsing==3.1.2
python==3.11.8
readline==8.2
requests==2.31.0
rsa=4.9
setuptools==68.2.2
sniffio==1.3.1
sqlite==3.41.2
starlette==0.37.2
tk==8.6.12
tqdm==4.66.2
typing-extensions==4.11.0
tzdata==2024a
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.41.2
xz==5.4.6
zlib==1.2.13